refactored element collection
collection to be replaced with individual collections
